requirements.txt
fastapi[standard]
sqlmodel
uvicorn
python-dotenv
psycopg2-binary
python-jose[cryptography] 
passlib[bcrypt]
pydantic_settings
alembic
testcontainers[postgresql]
pytest
faker